+++ SPRpy/SPRpy_functions.py
@@ -73,7 +73,7 @@ def load_csv_data(path=False, default_data_folder=None, prompt='Select the measu
     return data_path_, scanspeed, time_df, angles_df, ydata_df, reflectivity_df_
 
 # TODO: The new fitting parameters need to be supplied as pandas dataframes with each data point as a pandas series object
-def calculate_sensorgram(time, angles, ydata, TIR_range, scanspeed, SPR_boundary_points=(70, 70), SPR_fit_points=4000, TIR_window_count=None, TIR_fit_lower_ind=None, TIR_fit_higher_ind=None):
+def calculate_sensorgram(time, angles, ydata, TIR_range, scanspeed, SPR_boundary_points=(70, 70), SPR_fit_points=4000, TIR_window_count=None, TIR_fit_points=2000, TIR_fit_lower_ind=None, TIR_fit_higher_ind=None):
 
     # Convert dataframes to numpy ndarrays
     time = time.to_numpy()
@@ -82,22 +82,16 @@ def calculate_sensorgram(time, angles, ydata, TIR_range, scanspeed, SPR_boundary
 
     # TODO: The new fitting parameters need to be supplied as pandas dataframes with each data point as a pandas series object
     # Calculating SPR and TIR angles
-    sensorgram_SPR_angles = np.empty(len(ydata))
-    sensorgram_SPR_fit_y = np.empty(len(ydata))
-    sensorgram_SPR_fit_x = np.empty(len(ydata))
-    sensorgram_SPR_angles.fill(np.nan)
-    sensorgram_SPR_fit_x.fill(np.nan)
-    sensorgram_SPR_fit_y.fill(np.nan)
-    sensorgram_TIR_angles = np.empty(len(ydata))
-    sensorgram_TIR_deriv_x = np.empty(len(ydata))
-    sensorgram_TIR_deriv_y = np.empty(len(ydata))
-    sensorgram_TIR_deriv_fit_x = np.empty(len(ydata))
-    sensorgram_TIR_deriv_fit_y = np.empty(len(ydata))
-    sensorgram_TIR_angles.fill(np.nan)
-    sensorgram_TIR_deriv_x.fill(np.nan)
-    sensorgram_TIR_deriv_y.fill(np.nan)
-    sensorgram_TIR_deriv_fit_x.fill(np.nan)
-    sensorgram_TIR_deriv_fit_y.fill(np.nan)
+    sensorgram_SPR_angles = np.empty(len(ydata))*np.nan
+    sensorgram_SPR_fit_y = [pd.Series(np.empty(SPR_fit_points)*np.nan)]*len(ydata)
+    sensorgram_SPR_fit_x = [pd.Series(np.empty(SPR_fit_points)*np.nan)]*len(ydata)
+
+    sensorgram_TIR_angles = np.empty(len(ydata))*np.nan
+    TIR_deriv_points = len(angles[(angles >= TIR_range[0]) & (angles <= TIR_range[1])])
+    sensorgram_TIR_deriv_x = [pd.Series(np.empty(TIR_deriv_points)*np.nan)]*len(ydata)
+    sensorgram_TIR_deriv_y = [pd.Series(np.empty(TIR_deriv_points)*np.nan)]*len(ydata)
+    sensorgram_TIR_deriv_fit_x = [pd.Series(np.empty(TIR_fit_points)*np.nan)]*len(ydata)
+    sensorgram_TIR_deriv_fit_y = [pd.Series(np.empty(TIR_fit_points)*np.nan)]*len(ydata)
 
     for ind, val in enumerate(time):
         reflectivity_spectrum = ydata[ind-1, :]
@@ -115,31 +109,24 @@ def calculate_sensorgram(time, angles, ydata, TIR_range, scanspeed, SPR_boundary
             y_fit_min_ind = np.argmin(y_polyfit)
 
             sensorgram_SPR_angles[ind-1] = x_selection[y_fit_min_ind]
-            sensorgram_SPR_fit_x[ind-1] = x_selection
-            sensorgram_SPR_fit_y[ind-1] = y_polyfit
+            sensorgram_SPR_fit_x[ind-1] = pd.Series(x_selection)
+            sensorgram_SPR_fit_y[ind-1] = pd.Series(y_polyfit)
 
         except:
-            print('No SPR minimum found. Skipping measurement point...')
-            sensorgram_SPR_angles[ind-1] = np.nan
-            sensorgram_SPR_fit_x[ind-1] = np.nan
-            sensorgram_SPR_fit_y[ind-1] = np.nan
+            print('No SPR minimum found. Skipping measurement time point {}...'.format(val))
+
 
         # TIR angles
         try:
-            TIR_theta, TIR_xdata_filtered, deriv_ydata, TIR_theta_fit_x, TIR_theta_fit_y  = TIR_determination(angles, reflectivity_spectrum, TIR_range, scanspeed, _window_count=TIR_window_count, _fit_lower_ind=TIR_fit_lower_ind, _fit_higher_ind=TIR_fit_higher_ind)
+            TIR_theta, TIR_xdata_filtered, deriv_ydata, TIR_theta_fit_x, TIR_theta_fit_y  = TIR_determination(angles, reflectivity_spectrum, TIR_range, scanspeed, _window_count=TIR_window_count, _fit_points=TIR_fit_points, _fit_lower_ind=TIR_fit_lower_ind, _fit_higher_ind=TIR_fit_higher_ind)
             sensorgram_TIR_angles[ind-1] = TIR_theta
-            sensorgram_TIR_deriv_x[ind-1] = TIR_xdata_filtered
-            sensorgram_TIR_deriv_y[ind-1] = deriv_ydata
-            sensorgram_TIR_deriv_fit_x[ind-1] = TIR_theta_fit_x
-            sensorgram_TIR_deriv_fit_y[ind-1] = TIR_theta_fit_y
+            sensorgram_TIR_deriv_x[ind-1] = pd.Series(TIR_xdata_filtered)
+            sensorgram_TIR_deriv_y[ind-1] = pd.Series(deriv_ydata)
+            sensorgram_TIR_deriv_fit_x[ind-1] = pd.Series(TIR_theta_fit_x)
+            sensorgram_TIR_deriv_fit_y[ind-1] = pd.Series(TIR_theta_fit_y)
 
         except:
-            print('No TIR found. Skipping measurement point...')
-            sensorgram_TIR_angles[ind-1] = np.nan
-            sensorgram_TIR_deriv_x[ind - 1] = np.nan
-            sensorgram_TIR_deriv_y[ind - 1] = np.nan
-            sensorgram_TIR_deriv_fit_x[ind-1] = np.nan
-            sensorgram_TIR_deriv_fit_y[ind-1] = np.nan
+            print('No TIR found. Skipping measurement time point {}...'.format(val))
 
 
     sensorgram_df = pd.DataFrame(data={'time': time, 'SPR angle': sensorgram_SPR_angles, 'TIR angle': sensorgram_TIR_angles, 'SPR fit x': sensorgram_SPR_fit_x, 'SPR fit y': sensorgram_SPR_fit_y, 'TIR deriv x': sensorgram_TIR_deriv_x, 'TIR deriv y': sensorgram_TIR_deriv_y, 'TIR deriv fit x': sensorgram_TIR_deriv_fit_x, 'TIR deriv fit y': sensorgram_TIR_deriv_fit_y})