+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ### Features
 - Optical parameters for sensor tables are now loaded from values entered in "[default_sensor_values]" in config.toml
 - Can now change TIR and SPR angle fitting parameters for the sensorgram and refit them using a button hidden in a collapse
-- TODO: Displays the TIR and SPR angle fits in the reflectivity window when hovering over data points in the SPR sensorgram
+- Displays the TIR and SPR angle fits in the reflectivity window when hovering over data points in the SPR sensorgram
+- A custom folder location for the "SPRpy sessions" folder and for data prompt folder can now be set in config.toml 
 
 ### Fixes
--
+- Now includes time point information when TIR or SPR angle is not found.
 
 ## v1.0.1